pyproject.toml
[project]
name = "utig_radar_loading"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.25.0",
    "dask[complete]>=2025.9.1",
    "geoviews>=1.14.1",
    "h5py>=3.14.0",
    "hdf5storage>=0.2.1",
    "holoviews>=1.21.0",
    "hvplot>=0.12.1",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "utig-unfoc",
]

[tool.uv.workspace]
members = [
    "unfoc",
]

[tool.uv.sources]
utig-unfoc = { workspace = true }
